% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet10.1.R
\docType{data}
\name{DataSet10.1}
\alias{DataSet10.1}
\title{Data for Example 10.1 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Stroup, Ptukhina, and Garai (2024, 2nd ed.)}
\format{
A \code{data.frame} with 24 rows and 2 variables.
}
\usage{
data(DataSet10.1)
}
\description{
DataSet10.1 is used for One-way random effects model and BLUP estimation
}
\details{
\itemize{
       \item a  is a factor with 12 levels
       \item y  is a Gaussian response variable
       }
}
\examples{
data(DataSet10.1)
}
\references{
\enumerate{
\item Stroup, W. W., Ptukhina, M., and Garai, S. (2024).
     \emph{Generalized Linear Mixed Models: Modern Concepts, Methods and Applications (2nd ed.)}.
             CRC Press.
 }
}
\seealso{
\code{\link{Exam10.1}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
