% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet10.2.R
\docType{data}
\name{DataSet10.2}
\alias{DataSet10.2}
\title{Data for Example 10.2 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Stroup, Ptukhina, and Garai (2024, 2nd ed.)}
\format{
A \code{data.frame} with 28 rows and 3 variables with levels of b nested within levels of a.
}
\usage{
data(DataSet10.2)
}
\description{
DataSet10.2 is used for Two-way nested random effects model and BLUP estimation
}
\details{
\itemize{
       \item a  is a factor with 7 levels
       \item b  is a factor with 2 levels
       \item y  is a Gaussian response variable
       }
}
\examples{
data(DataSet10.2)
}
\references{
\enumerate{
\item Stroup, W. W., Ptukhina, M., and Garai, S. (2024).
     \emph{Generalized Linear Mixed Models: Modern Concepts, Methods and Applications (2nd ed.)}.
             CRC Press.
 }
}
\seealso{
\code{\link{Exam10.2}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
