% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet8.7.R
\docType{data}
\name{DataSet8.7}
\alias{DataSet8.7}
\title{Data for Example 8.7 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Stroup, Ptukhina, and Garai (2024, 2nd ed.)}
\usage{
data(DataSet8.7)
}
\description{
Data for Example 8.7 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Stroup, Ptukhina, and Garai (2024, 2nd ed.)
}
\examples{
data(DataSet8.7)
}
\references{
\enumerate{
\item Stroup, W. W., Ptukhina, M., and Garai, S. (2024).\emph{Generalized linear mixed models: modern concepts, methods and applications}.
             CRC Press.
 }
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Adeela Munawar (\email{adeela.uaf@gmail.com})
         }
}
\keyword{datasets}
