% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet9.3.R
\docType{data}
\name{DataSet9.3}
\alias{DataSet9.3}
\title{Data for Example 9.3 from Generalized Linear Mixed Models: Modern Concepts, Methods and Applications by Stroup, Ptukhina, and Garai (2024, 2nd ed.)}
\format{
A \code{data.frame} with 28 rows and 4 variables.
}
\usage{
data(DataSet9.3)
}
\description{
DataSet9.3 is used for Response surface design with incomplete blocking
}
\details{
\itemize{
       \item block  with 7 blocks
       \item a      is a factor with 3 levels 0,-1 and 1
       \item b      is a factor with 3 levels 0,-1 and 1
       \item c      is a factor with 3 levels 0,-1 and 1
       \item y      is a Gaussian response variable
       }
}
\examples{
data(DataSet9.3)
}
\references{
\enumerate{
\item Stroup, W. W., Ptukhina, M., and Garai, S. (2024).\emph{Generalized Linear Mixed Models: Modern Concepts, Methods and Applications (2nd ed.)}.
             CRC Press.
 }
}
\seealso{
\code{\link{Exam9.3}}
}
\author{
Muhammad Yaseen (\email{myaseen208@gmail.com})
        Adeela Munawar (\email{adeela.uaf@gmail.com})
}
\keyword{datasets}
